removed redundant old example code
